TChange all new test cases to use string spec instead
